requirements.txt
streamlit>=1.35.0
msal>=1.29.0
requests>=2.32.0
python-dotenv>=1.0.1